pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epog"
version = "0.1.0"
description = "Integrated Exploration and Sequential Manipulation on Scene Graph with LLM-based Situated Replanning"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = { text = "MIT" }
authors = [{ name = "Heqing Yang" }]
keywords = ["robotics", "scene-graph", "planning", "manipulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26.4,<2.0",
    "vedo",
    "trimesh",
    "opencv-python",
    "networkx",
    "transforms3d",
    "matplotlib",
    "tqdm",
    "typing-extensions",
    "pybullet",
    "python-fcl",
    "openai",
    "requests",
    "prettytable",
    "sentence-transformers",
    "prior",
    "ai2thor-colab",
    "manifold3d",
    "python-tsp",
    "loguru",
    "python-astar",
    "pandas",
    "sdf",
]

[dependency-groups]
dev = [
    "pre-commit>=3.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "types-requests",
    "pandas-stubs",
]

[project.urls]
Homepage = "https://github.com/yangheqing/epog"
Repository = "https://github.com/yangheqing/epog"

[tool.uv.sources]
sdf = { path = "third_party/sdf", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["pog", "epog"]

# Ruff configuration (replaces black, isort, flake8)
[tool.ruff]
target-version = "py310"
line-length = 88
exclude = ["third_party/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument
    "F402",   # import shadowed by loop variable
]

[tool.ruff.lint.isort]
known-first-party = ["epog", "pog"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["third_party/"]

[[tool.mypy.overrides]]
module = ["vedo.*", "trimesh.*", "pybullet.*", "fcl.*", "ai2thor.*", "prior.*"]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
